+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## next
 
+### Features
+
+- new rules `id-pattern` and `class-pattern` for requiring a specific formats.
+- new rule `no-dup-class` preventing duplicate classes names on the same
+  element.
+
 ## 0.2.0 (2017-10-11)
 
 ### Features